Remove some meteor deep imports
