+++ apps/meteor/app/lib/lib/MessageTypesValues.ts
@@ -1,0 +1,136 @@
+import type { MessageTypesValues as MessageTypesValuesType } from '@rocket.chat/core-typings';
+
+export const MessageTypesValues: Array<{ key: MessageTypesValuesType; i18nLabel: string }> = [
+	{
+		key: 'uj', // user joined
+		i18nLabel: 'Message_HideType_uj',
+	},
+	{
+		key: 'ujt', // user joined team
+		i18nLabel: 'Message_HideType_ujt',
+	},
+	{
+		key: 'ul', // user left
+		i18nLabel: 'Message_HideType_ul',
+	},
+	{
+		key: 'ult', // user left team
+		i18nLabel: 'Message_HideType_ult',
+	},
+	{
+		key: 'ru', // user removed
+		i18nLabel: 'Message_HideType_ru',
+	},
+	{
+		key: 'removed-user-from-team',
+		i18nLabel: 'Message_HideType_removed_user_from_team',
+	},
+	{
+		key: 'au', // added user
+		i18nLabel: 'Message_HideType_au',
+	},
+	{
+		key: 'added-user-to-team',
+		i18nLabel: 'Message_HideType_added_user_to_team',
+	},
+	{
+		key: 'mute_unmute',
+		i18nLabel: 'Message_HideType_mute_unmute',
+	},
+	{
+		key: 'r', // room name changed
+		i18nLabel: 'Message_HideType_r',
+	},
+	{
+		key: 'ut', // user joined conversation
+		i18nLabel: 'Message_HideType_ut',
+	},
+	{
+		key: 'wm', // welcome
+		i18nLabel: 'Message_HideType_wm',
+	},
+	{
+		key: 'rm', // message removed
+		i18nLabel: 'Message_HideType_rm',
+	},
+	{
+		key: 'subscription-role-added',
+		i18nLabel: 'Message_HideType_subscription_role_added',
+	},
+	{
+		key: 'subscription-role-removed',
+		i18nLabel: 'Message_HideType_subscription_role_removed',
+	},
+	{
+		key: 'room-archived',
+		i18nLabel: 'Message_HideType_room_archived',
+	},
+	{
+		key: 'room-unarchived',
+		i18nLabel: 'Message_HideType_room_unarchived',
+	},
+	{
+		key: 'room_changed_privacy',
+		i18nLabel: 'Message_HideType_room_changed_privacy',
+	},
+	{
+		key: 'room_changed_avatar',
+		i18nLabel: 'Message_HideType_room_changed_avatar',
+	},
+	{
+		key: 'room_changed_topic',
+		i18nLabel: 'Message_HideType_room_changed_topic',
+	},
+	{
+		key: 'room_e2e_enabled',
+		i18nLabel: 'Message_HideType_room_enabled_encryption',
+	},
+	{
+		key: 'room_e2e_disabled',
+		i18nLabel: 'Message_HideType_room_disabled_encryption',
+	},
+	{
+		key: 'room-removed-read-only',
+		i18nLabel: 'Message_HideType_room_removed_read_only',
+	},
+	{
+		key: 'room-set-read-only',
+		i18nLabel: 'Message_HideType_room_set_read_only',
+	},
+	{
+		key: 'room-disallowed-reacting',
+		i18nLabel: 'Message_HideType_room_disallowed_reacting',
+	},
+	{
+		key: 'room-allowed-reacting',
+		i18nLabel: 'Message_HideType_room_allowed_reacting',
+	},
+	{
+		key: 'user-added-room-to-team',
+		i18nLabel: 'Message_HideType_user_added_room_to_team',
+	},
+	{
+		key: 'user-converted-to-channel',
+		i18nLabel: 'Message_HideType_user_converted_to_channel',
+	},
+	{
+		key: 'user-converted-to-team',
+		i18nLabel: 'Message_HideType_user_converted_to_team',
+	},
+	{
+		key: 'user-deleted-room-from-team',
+		i18nLabel: 'Message_HideType_user_deleted_room_from_team',
+	},
+	{
+		key: 'user-removed-room-from-team',
+		i18nLabel: 'Message_HideType_user_removed_room_from_team',
+	},
+	{
+		key: 'room_changed_announcement',
+		i18nLabel: 'Message_HideType_changed_announcement',
+	},
+	{
+		key: 'room_changed_description',
+		i18nLabel: 'Message_HideType_changed_description',
+	},
+];
